Rename file to not include factory
