+++ arches_her/views/file_template.py
@@ -43,7 +43,7 @@ from arches.app.models.tile import Tile
 from arches.app.utils.response import JSONResponse
 from arches.app.views.tile import TileData
 import sysconfig
-from arches_her.utils.file_template_modifier_factory import FileTemplateModifierFactory
+from arches_her.utils.file_template_modifier import FileTemplateModifierFactory
 
 class FileTemplateView(View):
     def __init__(self):
